requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
chromadb==0.5.3
sentence-transformers==3.2.0
httpx==0.27.2
python-dotenv==1.0.1
werkzeug==3.0.1
PyPDF2==3.0.1
jinja2==3.1.4
starlette==0.40.0
jinja_partials
fastapi-websocket-pubsub
markdown
python-multipart